+++ packages/digit_data_model/lib/data/repositories/oplog/oplog.dart
@@ -144,7 +144,7 @@ abstract class OpLogManager<T extends EntityModel> {
       });
     } catch (e) {
       if (kDebugMode) {
-        print('error in isar ${e}');
+        print('error in isar $e');
       }
       rethrow;
     }