+++ packages/registration_delivery/lib/data/repositories/local/household_global_search.dart
@@ -497,7 +497,7 @@ class HouseHoldGlobalSearchRepository extends LocalRepository {
     var variables = selectQuery.constructQuery().introducedVariables;
     var indexesLength = selectQuery.constructQuery().variableIndices;
 
-    var totalCount;
+    dynamic totalCount;
 
     try {
       totalCount = await sql