+++ packages/sync_service/test/sync_service_test.dart
@@ -1,7 +1,3 @@
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:sync_service/sync_service_lib.dart';
-
 void main() {
 
 }